feat: ChessGame 컨트롤러로 분리
