+++ src/main/java/chess/domain/Board.java
@@ -63,8 +63,6 @@ public class Board {
         }
     }
 
-//    public void load()
-
     public ChessGameResult calculateGameScore() {
         return scoreCalculator.calculate(pieces);
     }
